--- a/docker/c7n
+++ b/docker/c7n
@@ -88,10 +88,6 @@ FROM ubuntu:22.04
 LABEL name="cli" \
       repository="http://github.com/cloud-custodian/cloud-custodian"
 
-COPY --from=build-env /src /src
-COPY --from=build-env /usr/local /usr/local
-COPY --from=build-env /output /output
-
 ENV DEBIAN_FRONTEND=noninteractive
 
 RUN apt-get --yes update \
@@ -99,6 +95,13 @@ RUN apt-get --yes update \
         && rm -Rf /var/cache/apt \
         && rm -Rf /var/lib/apt/lists/* \
         && rm -Rf /var/log/*
+
+# These should remain below any other commands because they will invalidate
+# the layer cache
+COPY --from=build-env /src /src
+COPY --from=build-env /usr/local /usr/local
+COPY --from=build-env /output /output
+
 
 RUN adduser --disabled-login --gecos "" custodian
 USER custodian