--- a/docker/c7n-org
+++ b/docker/c7n-org
@@ -3,6 +3,7 @@
 FROM ubuntu:22.04 as build-env
 
 ARG POETRY_VERSION="1.2.1"
+SHELL ["/bin/bash", "-c"]
 
 # pre-requisite distro deps, and build env setup
 RUN adduser --disabled-login --gecos "" custodian
@@ -15,29 +16,70 @@ WORKDIR /src
 
 # Add core & aws packages
 ADD pyproject.toml poetry.lock README.md /src/
-ADD c7n /src/c7n/
 RUN . /usr/local/bin/activate && pip install -U pip
-RUN . /usr/local/bin/activate && poetry install --no-dev
+
+# Ignore root first pass so if source changes we don't have to invalidate
+# dependency install
+RUN . /usr/local/bin/activate && poetry install --without dev --no-root
 RUN . /usr/local/bin/activate && pip install -q wheel &&       pip install -U pip
 RUN . /usr/local/bin/activate && pip install -q aws-xray-sdk psutil jsonpatch
 
+ARG providers="gcp azure kube openstack tencentcloud"
 # Add provider packages
-ADD tools/c7n_gcp /src/tools/c7n_gcp
-RUN rm -R tools/c7n_gcp/tests
-ADD tools/c7n_azure /src/tools/c7n_azure
-RUN rm -R tools/c7n_azure/tests_azure
-ADD tools/c7n_kube /src/tools/c7n_kube
-RUN rm -R tools/c7n_kube/tests
-ADD tools/c7n_openstack /src/tools/c7n_openstack
-RUN rm -R tools/c7n_openstack/tests
-ADD tools/c7n_tencentcloud /src/tools/c7n_tencentcloud
-RUN rm -R tools/c7n_tencentcloud/tests
+# We include `pyproject.toml` and `poetry.lock` first to allow
+# cache of dependency installs.
 
-# Install requested providers
-ARG providers="gcp kube openstack tencentcloud azure"
-RUN . /usr/local/bin/activate && \
-    for pkg in $providers; do cd tools/c7n_$pkg && \
-    poetry install && cd ../../; done
+ADD tools/c7n_gcp/pyproject.toml tools/c7n_gcp/poetry.lock /src/tools/c7n_gcp/
+RUN if [[ " ${providers[*]} " =~ " gcp " ]]; then     . /usr/local/bin/activate &&     cd tools/c7n_gcp &&     poetry install --without dev --no-root; fi
+
+
+ADD tools/c7n_azure/pyproject.toml tools/c7n_azure/poetry.lock /src/tools/c7n_azure/
+RUN if [[ " ${providers[*]} " =~ " azure " ]]; then     . /usr/local/bin/activate &&     cd tools/c7n_azure &&     poetry install --without dev --no-root; fi
+
+
+ADD tools/c7n_kube/pyproject.toml tools/c7n_kube/poetry.lock /src/tools/c7n_kube/
+RUN if [[ " ${providers[*]} " =~ " kube " ]]; then     . /usr/local/bin/activate &&     cd tools/c7n_kube &&     poetry install --without dev --no-root; fi
+
+
+ADD tools/c7n_openstack/pyproject.toml tools/c7n_openstack/poetry.lock /src/tools/c7n_openstack/
+RUN if [[ " ${providers[*]} " =~ " openstack " ]]; then     . /usr/local/bin/activate &&     cd tools/c7n_openstack &&     poetry install --without dev --no-root; fi
+
+
+ADD tools/c7n_tencentcloud/pyproject.toml tools/c7n_tencentcloud/poetry.lock /src/tools/c7n_tencentcloud/
+RUN if [[ " ${providers[*]} " =~ " tencentcloud " ]]; then     . /usr/local/bin/activate &&     cd tools/c7n_tencentcloud &&     poetry install --without dev --no-root; fi
+
+
+# Now install the root package
+ADD c7n /src/c7n/
+RUN . /usr/local/bin/activate && poetry install --only-root
+
+# Now install the root of each provider
+
+ADD tools/c7n_gcp /src/tools/c7n_gcp
+RUN if [[ " ${providers[*]} " =~ " gcp " ]]; then     . /usr/local/bin/activate &&     cd tools/c7n_gcp &&     poetry install --only-root; fi
+
+
+ADD tools/c7n_gcp /src/tools/c7n_gcp
+RUN if [[ " ${providers[*]} " =~ " azure " ]]; then     . /usr/local/bin/activate &&     cd tools/c7n_azure &&     poetry install --only-root; fi
+
+
+ADD tools/c7n_gcp /src/tools/c7n_gcp
+RUN if [[ " ${providers[*]} " =~ " kube " ]]; then     . /usr/local/bin/activate &&     cd tools/c7n_kube &&     poetry install --only-root; fi
+
+
+ADD tools/c7n_gcp /src/tools/c7n_gcp
+RUN if [[ " ${providers[*]} " =~ " openstack " ]]; then     . /usr/local/bin/activate &&     cd tools/c7n_openstack &&     poetry install --only-root; fi
+
+
+ADD tools/c7n_gcp /src/tools/c7n_gcp
+RUN if [[ " ${providers[*]} " =~ " tencentcloud " ]]; then     . /usr/local/bin/activate &&     cd tools/c7n_tencentcloud &&     poetry install --only-root; fi
+
+
+ADD tools/c7n_gcp /src/tools/c7n_gcp
+ADD tools/c7n_azure /src/tools/c7n_azure
+ADD tools/c7n_kube /src/tools/c7n_kube
+ADD tools/c7n_openstack /src/tools/c7n_openstack
+ADD tools/c7n_tencentcloud /src/tools/c7n_tencentcloud
 
 RUN mkdir /output