--- a/docker/c7n-org
+++ b/docker/c7n-org
@@ -75,12 +75,6 @@ ADD tools/c7n_tencentcloud /src/tools/c7n_tencentcloud
 RUN if [[ " ${providers[*]} " =~ "tencentcloud" ]]; then     . /usr/local/bin/activate &&     cd tools/c7n_tencentcloud &&     poetry install --only-root; fi
 
 
-ADD tools/c7n_gcp /src/tools/c7n_gcp
-ADD tools/c7n_azure /src/tools/c7n_azure
-ADD tools/c7n_kube /src/tools/c7n_kube
-ADD tools/c7n_openstack /src/tools/c7n_openstack
-ADD tools/c7n_tencentcloud /src/tools/c7n_tencentcloud
-
 RUN mkdir /output
 
 # Install c7n-org